extract stardust wallet into its own module
